requirements.txt
Django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
psycopg2-binary==2.9.9
cryptography==42.0.8
Pillow==10.3.0
requests==2.32.3

celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0

# Then run:
#   pip install celery redis django-celery-beat --break-system-packages
#   python manage.py migrate django_celery_beat